added player salaries

+++ roster/models.py
@@ -9,8 +9,8 @@ class Player(models.Model):
     first_name = models.CharField(max_length=50)
     last_name = models.CharField(max_length=50)
     position = models.CharField(max_length=2, choices=POSITION)
-    cap_hit = models.IntegerField()
-    cut_savings = models.IntegerField(null=True,blank=True)
+    cap_hit = models.FloatField()
+    cut_savings = models.FloatField(null=True,blank=True)
 
     def __str__(self):
         return str(self.position + ' ' + self.first_name + " " + self.last_name)
@@ -19,7 +19,7 @@ class FreeAgent(models.Model):
     first_name = models.CharField(max_length=50)
     last_name = models.CharField(max_length=50)
     position = models.CharField(max_length=2, choices=POSITION)
-    cap_hit = models.IntegerField()
+    cap_hit = models.FloatField()
     on_team = models.BooleanField()
 
     def __str__(self):
@@ -29,7 +29,7 @@ class DraftPlayer(models.Model):
     first_name = models.CharField(max_length=50)
     last_name = models.CharField(max_length=50)
     position = models.CharField(max_length=2, choices=POSITION)
-    cap_hit = models.IntegerField()
+    cap_hit = models.FloatField()
     draft_round = models.IntegerField()
     college = models.CharField(max_length=100)
